# Various packaging and performance fixes