--- a/crc32c.cabal
+++ b/crc32c.cabal
@@ -76,3 +76,14 @@ test-suite crc32c-test
     , hspec
     , hspec-core
   default-language: Haskell2010
+
+benchmark crc32c-bench
+  type: exitcode-stdio-1.0
+  main-is: Bench.hs
+  hs-source-dirs: bench
+  build-depends:
+      base
+    , bytestring
+    , crc32c
+    , criterion
+  default-language: Haskell2010