--- a/crc32c.cabal
+++ b/crc32c.cabal
@@ -54,12 +54,12 @@ library
     , bytestring >=0.10.8.2
   default-language: Haskell2010
   if arch(x86_64)
-    cxx-options: -std=c++11 -D__HAVE_SSE42=1 -D__HAVE_ARM64_CRC32C=0 -msse4.2
+    cxx-options: -std=c++11 -DHAVE_SSE42=1 -DHAVE_ARM64_CRC32C=0 -msse4.2
   else
     if arch(arm)
-      cxx-options: -std=c++11 -D__HAVE_SSE42=0 -D__HAVE_ARM64_CRC32C=1
+      cxx-options: -std=c++11 -DHAVE_SSE42=0 -DHAVE_ARM64_CRC32C=1
     else
-      cxx-options: -std=c++11 -D__HAVE_SSE42=0 -D__HAVE_ARM64_CRC32C=0
+      cxx-options: -std=c++11 -DHAVE_SSE42=0 -DHAVE_ARM64_CRC32C=0
 
 test-suite crc32c-test
   type: exitcode-stdio-1.0