--- a/crc32c.cabal
+++ b/crc32c.cabal
@@ -34,8 +34,6 @@ extra-source-files:
 library
   exposed-modules:
       Data.Digest.CRC32C
-  other-modules:
-      Data.Digest.CRC32C.Internal
   hs-source-dirs:
       src
   include-dirs:
@@ -47,8 +45,6 @@ library
       include/crc32c_sse42.cc
   extra-libraries:
       stdc++
-  build-tools:
-      c2hs
   build-depends:
       base >=4.9 && <5
     , bytestring >=0.10.8.2