--- a/crc32c.cabal
+++ b/crc32c.cabal
@@ -64,8 +64,8 @@ test-suite crc32c-test
       test
   build-depends:
       QuickCheck
-    , base >=4.9 && <5
-    , bytestring >=0.10.8.2 && <0.12
+    , base
+    , bytestring
     , crc32c
     , hspec
     , hspec-core