--- a/src/Data/Digest/CRC32C.hs
+++ b/src/Data/Digest/CRC32C.hs
@@ -4,7 +4,6 @@ module Data.Digest.CRC32C(
   ) where
 
 import           Data.ByteString.Internal    (ByteString (..))
-import           Data.Digest.CRC32C.Internal
 import           Data.Word
 import           Foreign.C.Types
 import           Foreign.ForeignPtr.Unsafe
@@ -16,3 +15,8 @@ crc32c (PS p o l) = fromIntegral $ lib_crc32c_value (unsafeForeignPtrToPtr p `pl
 crc32c_update :: Word32 -> ByteString -> Word32
 crc32c_update hash (PS p o l) = fromIntegral $ lib_crc32c_extend (CUInt hash) (unsafeForeignPtrToPtr p `plusPtr` o) (fromIntegral l)
 
+foreign import ccall "crc32c/crc32c.h crc32c_extend"
+  lib_crc32c_extend :: CUInt -> Ptr CUChar -> CSize -> CUInt
+
+foreign import ccall "crc32c/crc32c.h crc32c_value"
+  lib_crc32c_value :: Ptr CUChar -> CSize -> CUInt