… src/Data/Digest/CRC32C/Internal.chs