We start by adding a simple benchmark.

Here are the baseline results on my box, before any fixes and improvements.
```
benchmarking crc32c small
time                 419.6 ns   (416.5 ns .. 423.4 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 419.1 ns   (417.2 ns .. 423.1 ns)
std dev              9.071 ns   (5.273 ns .. 16.31 ns)
variance introduced by outliers: 28% (moderately inflated)

benchmarking crc32c large
time                 59.44 μs   (59.17 μs .. 59.69 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 59.33 μs   (59.16 μs .. 59.62 μs)
std dev              746.9 ns   (564.4 ns .. 1.049 μs)
```
where "small" is 2^8 bytes and large is 2^16 bytes.

The first fix is then to compile the C++ code properly, using `cxx-sources` and `cxx-options`. This already improves the benchmarks substantially:
```
benchmarking crc32c small
time                 204.0 ns   (203.2 ns .. 205.0 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 204.3 ns   (203.6 ns .. 204.9 ns)
std dev              2.098 ns   (1.765 ns .. 2.504 ns)

benchmarking crc32c large
time                 16.99 μs   (16.94 μs .. 17.03 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 16.97 μs   (16.92 μs .. 17.02 μs)
std dev              164.7 ns   (126.0 ns .. 236.3 ns)
```
Note that this fix should also fix building on OSX, by passing `-std=c++11`.

Next is to use the specialised amd64 SSE4.2 and arm implementations on the supported platforms, rather than the generic version, by using the correct CPP defines (without the __ prefixes). This also improves the benchmarks substantially.
```
benchmarking crc32c small
time                 129.5 ns   (129.1 ns .. 130.0 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 130.3 ns   (129.9 ns .. 130.7 ns)
std dev              1.417 ns   (1.050 ns .. 1.894 ns)

benchmarking crc32c large
time                 2.607 μs   (2.602 μs .. 2.611 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.603 μs   (2.597 μs .. 2.617 μs)
std dev              31.05 ns   (16.13 ns .. 66.94 ns)
```