@dcoutts 2024-02-09
This is a WIP intended to address #3.