@dcoutts 2024-02-12
Final version, after using `unsafe` FFI calls is around 15x faster for the small case (2^8 bytes) and around 25x faster for the large case (2^16 bytes).
```
benchmarking crc32c small
time                 26.76 ns   (26.57 ns .. 26.98 ns)
                     1.000 R²   (0.999 R² .. 1.000 R²)
mean                 26.79 ns   (26.65 ns .. 26.95 ns)
std dev              486.5 ps   (405.3 ps .. 631.3 ps)
variance introduced by outliers: 25% (moderately inflated)

benchmarking crc32c large
time                 2.345 μs   (2.332 μs .. 2.357 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.345 μs   (2.335 μs .. 2.356 μs)
std dev              36.49 ns   (29.70 ns .. 49.10 ns)
variance introduced by outliers: 15% (moderately inflated)
```