@dcoutts 2024-02-12
This should now also build out of the box on Windows and OSX, fixing #3.