@dcoutts 2024-02-12
Note this is split into lots of little patches with the intention of making review easier. If you prefer it all squashed together, let me know.