@dcoutts 2024-02-12
At the suggestion of @phadej, use safe or unsafe FFI below/above a size threshold to avoid long running calls blocking other threads or GC.

I picked the threshold to be 16k, and I've extended the benchmark so you can see the effect: 16k+1 takes about 8% longer than 16k.
```
benchmarking crc32c small (2^8)
time                 25.24 ns   (25.18 ns .. 25.29 ns)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 25.21 ns   (25.13 ns .. 25.29 ns)
std dev              274.3 ps   (210.2 ps .. 365.5 ps)
variance introduced by outliers: 11% (moderately inflated)

benchmarking crc32c large (2^16)
time                 2.284 μs   (2.276 μs .. 2.293 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.283 μs   (2.277 μs .. 2.288 μs)
std dev              18.42 ns   (13.74 ns .. 24.63 ns)

benchmarking crc32c large (2^16+1)
time                 2.469 μs   (2.458 μs .. 2.481 μs)
                     1.000 R²   (1.000 R² .. 1.000 R²)
mean                 2.471 μs   (2.463 μs .. 2.478 μs)
std dev              26.60 ns   (20.54 ns .. 36.22 ns)
```